provisional: heuristic improvement for RotatingBloomFilter
I would expect that (at least for Moz applications) the objects would
tend to cluster. In other words that propability to see old object
is lower than to see the young one. It does not matter for the classic
bloom filter, but the proposed change would improve performance of the
rotating one.